+++ Assets/UniGLTF/Editor/Animation/AnimationCurveData.cs
