+++ Assets/UniGLTF/Editor/Animation/AnimationCurveData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cd3120ad83927864fa7b56ce619e60ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
