+++ Assets/UniGLTF/Editor/Animation/AnimationExporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 10bbffd95db4d7a4abb3d0e74f46f994
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
