+++ Assets/UniGLTF/Editor/Animation/AnimationKeyframeData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 343050b608326344ebef0369448f8159
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
