+++ Assets/UniGLTF/Editor/Animation/EditorAnimationExporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 62c5fa8e50275804b8084f3764f9d8f3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
