--- Assets/UniGLTF/Runtime/UniGLTF/IO/AnimationIO/AnimationCurveData.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1af21722605d44f58deebfcfca642b32
-timeCreated: 1537442711
