--- Assets/UniGLTF/Runtime/UniGLTF/IO/AnimationIO/AnimationExporter.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 015ae41bf6cb4428b8257ead79772908
-timeCreated: 1537443293
