--- Assets/UniGLTF/Runtime/UniGLTF/IO/AnimationIO/AnimationKeyframeData.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: e27ef4fb768e49f591c2bb5eadd3b19b
-timeCreated: 1537442737
