+++ Assets/UniGLTF/Runtime/UniGLTF/IO/AnimationIO/IAnimationExporter.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace UniGLTF
+{
+    public interface IAnimationExporter
+    {
+        void Export(ExportingGltfData _data, GameObject Copy, List<Transform> Nodes);
+    }
+}
