+++ Assets/UniGLTF/Runtime/UniGLTF/IO/AnimationIO/IAnimationExporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7bcb89e584f47c34c9ad19bdce8ccad0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
