+++ Assets/UniGLTF/Runtime/UniGLTF/IO/GltfExportSettings.cs
@@ -45,5 +45,7 @@ namespace UniGLTF
         /// Keep VertexColor
         /// </summary>
         public bool KeepVertexColor;
+
+        public IAnimationExporter AnimationExporter;
     }
 }
